+++ @xen-orchestra/lite/src/components/LinkWrapper.tsx
@@ -1,18 +1,8 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-const LinkWrapper = ({
-  children,
-  link,
-  to,
-  style,
-}: {
-  children: React.ReactNode
-  link?: boolean
-  to?: string | object
-  style?: object
-}) =>
-  link ? (
+const LinkWrapper = ({ children, to, style }: { children: React.ReactNode; to?: string | object; style?: object }) =>
+  to !== undefined ? (
     <Link to={to} style={style}>
       {children}
     </Link>
